Bug 3 fixed
Fixed bugs 3a and 3b, resulting in bug 3 being fixed. Added the resolution report.Edited the automated test to be more strict in account of the sheer number of trials that are being ran. Fixed typos in v0.1 of Test Case 3, added v1.0 of Test Case 3, and edited the debugging log? I must not have saved it prior to the previous commit.